pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "s1s2"
version = "0.1.0"
description = "Mechanistic signatures of dual-process cognition in LLMs"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "HUSAI S1/S2 Team" },
]
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "torch>=2.4",
    "transformers>=4.45",
    "accelerate>=0.34",
    "numpy>=1.26",
    "scipy>=1.13",
    "scikit-learn>=1.5",
    "pandas>=2.2",
    "h5py>=3.11",
    "matplotlib>=3.9",
    "seaborn>=0.13",
    "umap-learn>=0.5.6",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "beartype>=0.18",
    "jaxtyping>=0.2.34",
    "einops>=0.8",
    "wandb>=0.18",
    "statsmodels>=0.14",
    "tqdm>=4.66",
    "datasets>=3.0",
    "pyyaml>=6.0",
    "rich>=13.7",
]

[project.optional-dependencies]
sae = [
    "sae-lens>=4.0",
    "transformer-lens>=2.5",
    "nnsight>=0.3",
]
dev = [
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "black>=24.8",
    "mypy>=1.11",
    "pre-commit>=3.8",
    "ipykernel>=6.29",
    "jupyterlab>=4.2",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["s1s2*"]

[tool.setuptools.package-data]
s1s2 = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["notebooks", "data", "results", "figures"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "RUF", # ruff-specific
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults (Hydra)
]

[tool.black]
line-length = 100
target-version = ["py311"]
extend-exclude = "notebooks/|data/|results/|figures/"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --strict-markers --tb=short"
markers = [
    "slow: marks tests as slow (deselect with -m 'not slow')",
    "gpu: marks tests requiring GPU",
    "integration: marks integration tests",
]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
plugins = ["numpy.typing.mypy_plugin"]
